[core] Fix deadlock when cancelling stale requests on in-order actors (#57746)
