+++ src/ray/core_worker/task_execution/task_receiver.h
@@ -104,10 +104,8 @@ class TaskReceiver {
   void SetActorReprName(const std::string &repr_name);
 
  private:
-  /// Guard for shutdown state.
-  absl::Mutex stop_mu_;
   // True once shutdown begins. Requests to execute new tasks will be rejected.
-  bool stopping_ ABSL_GUARDED_BY(stop_mu_) = false;
+  std::atomic<bool> stopping_ = false;
   /// Set up the configs for an actor.
   /// This should be called once for the actor creation task.
   void SetupActor(bool is_asyncio,
